[OPS-5828][OPS-5687] Patch all (possibly) used versions of path.inc.
Redis module has its own path.inc which should be used *everywhere* by
core. However, our core path.inc is patched by path_alias_xt module and
this means that on order to get reliable operation, we *also* need to
patch the redis-provided path.inc version with the path_alias_xt fix.

The alternative is to install PECL runkit, but that has security
implications we would rather not deal with.